refactor: restrict visibility of gadgets::utils
